+++ Dtos/User/GetUserDto.cs
@@ -1,5 +1,5 @@
 using System.ComponentModel.DataAnnotations;
-using MediaTrackerAuthenticationService.Models.Utils;
+using MediaTrackerAuthenticationService.Models.AuthDB;
 
 namespace MediaTrackerAuthenticationService.Dtos.User
 {
@@ -7,7 +7,6 @@ namespace MediaTrackerAuthenticationService.Dtos.User
     {
         public required int  UserId   { get; set; }
         public required MediaPlatform Platform { get; set; }
-
         public required string PlatformId { get; set; }
     }
 }
